chg: rename the _abstract module into a _visual module as it contains only the Visual class
